+++ pyproject.toml
@@ -12,6 +12,7 @@ dependencies = [
     "click>=8.2.1",
     "openpyxl>=3.1.5",
     "flask>=3.1.2",
+    "jinja2>=3.1.6",
 ]
 
 [project.scripts]
